+++ packages/babel-preset-udemy-website/tests/result/css-modules.js
@@ -17,27 +17,29 @@ var CssModules = function (_Component) {
 
     function CssModules() {
         babelHelpers.classCallCheck(this, CssModules);
-        return babelHelpers.possibleConstructorReturn(this, _Component.apply(this, arguments));
+        return babelHelpers.possibleConstructorReturn(this, (CssModules.__proto__ || Object.getPrototypeOf(CssModules)).apply(this, arguments));
     }
 
-    CssModules.prototype.render = function render() {
-        // styles.greeting can be evaluated at compile time.
-        // baseStyles.colorful won't actually work at runtime since the *.global.less file is
-        // ignored, but I'm testing that it doesn't try to actually output a lookup table
-        // for baseStyles as it will for styles since I'm doing a dynamic styleName.
-        return React.createElement(
-            'div',
-            { className: 'colorful css-modules--greeting--3-JKt' },
-            React.createElement(
-                'span',
-                {
-                    className: _getClassName('baseStyles.colorful', _styleModuleImportMap)
-                },
-                'Hello World!'
-            )
-        );
-    };
-
+    babelHelpers.createClass(CssModules, [{
+        key: 'render',
+        value: function render() {
+            // styles.greeting can be evaluated at compile time.
+            // baseStyles.colorful won't actually work at runtime since the *.global.less file is
+            // ignored, but I'm testing that it doesn't try to actually output a lookup table
+            // for baseStyles as it will for styles since I'm doing a dynamic styleName.
+            return React.createElement(
+                'div',
+                { className: 'colorful css-modules--greeting--3-JKt' },
+                React.createElement(
+                    'span',
+                    {
+                        className: _getClassName('baseStyles.colorful', _styleModuleImportMap)
+                    },
+                    'Hello World!'
+                )
+            );
+        }
+    }]);
     return CssModules;
 }(Component);