+++ packages/babel-preset-udemy-website/tests/result/for-of-destructuring.js
@@ -3,22 +3,33 @@
 var array = new Array(1e3).fill(1);
 var counter = 0;
 
-for (var _iterator = array.entries(), _isArray = Array.isArray(_iterator), _i = 0, _iterator = _isArray ? _iterator : _iterator[Symbol.iterator]();;) {
-  var _ref;
+var _iteratorNormalCompletion = true;
+var _didIteratorError = false;
+var _iteratorError = undefined;
 
-  if (_isArray) {
-    if (_i >= _iterator.length) break;
-    _ref = _iterator[_i++];
-  } else {
-    _i = _iterator.next();
-    if (_i.done) break;
-    _ref = _i.value;
+try {
+  for (var _iterator = array.entries()[Symbol.iterator](), _step; !(_iteratorNormalCompletion = (_step = _iterator.next()).done); _iteratorNormalCompletion = true) {
+    var _ref = _step.value;
+
+    var _ref2 = babelHelpers.slicedToArray(_ref, 2);
+
+    var i = _ref2[0];
+    var val = _ref2[1];
+    counter += i + val;
   }
+} catch (err) {
+  _didIteratorError = true;
+  _iteratorError = err;
+} finally {
+  try {
+    if (!_iteratorNormalCompletion && _iterator.return) {
+      _iterator.return();
+    }
+  } finally {
+    if (_didIteratorError) {
+      throw _iteratorError;
+    }
+  }
+}
 
-  var _ref2 = _ref,
-      _ref3 = babelHelpers.slicedToArray(_ref2, 2),
-      i = _ref3[0],
-      val = _ref3[1];
-
-  counter += i + val;
-}module.exports = counter;
+module.exports = counter;